backend/covalent/requirements.txt
pymongo==3.11.0
pytz==2020.1
python-dateutil==2.8.1